+++ src/utils/exam.ts
@@ -2,6 +2,7 @@ import { GENERAL_CATEGORY_ID } from '../constants'
 import categories from '../data/exam-data/categories.json'
 import examTypes from '../data/exam-data/examTypes.json'
 import type { Exam, GeneratedExam, Question, LangCode, ExamType } from '../types'
+import { randomizeChoices, shuffleArray } from './format'
 
 // Map to retrieve questions in order later (better performance O(M + N) instead of O(M * N))
 // M is size of questions in exam, N is size of question bank
@@ -25,20 +26,6 @@ export async function initQuestionMap(langCode: LangCode) {
         questionMap = new Map();
         throw new Error(`Couldn't open file: 'questions-${langCode}.json'`)
     }
-}
-
-/**
-   * Shuffle array using Fisher-Yates algorithm
-   * @param {T[]} array - The array to shuffle
-   * @returns {T[]} - The shuffled array
-   */
-function shuffleArray<T>(array: T[]): T[] {
-    const shuffled = [...array]
-    for (let i = shuffled.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1))
-            ;[shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]]
-    }
-    return shuffled
 }
 
 /**
@@ -123,7 +110,7 @@ export function generateNewExam(examType: ExamType, categoryId: number): Generat
     const questionIds = chosenQuestions.map((q: Question): Question['id'] => q.id)
 
     const generatedExam: GeneratedExam = {
-        exam: chosenQuestions,
+        exam: randomizeChoices(chosenQuestions),
         questionIds,
         durationMinutes
     }
